[202] Update the way change proposals are displayed
Bug: https://github.com/svalyn/svalyn-studio/issues/202
