+++ frontend/svalyn-studio-app/src/projects/ProjectViewHeader.types.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022 Stéphane Bégaudeau.
+ * Copyright (c) 2023 Stéphane Bégaudeau.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
  * associated documentation files (the "Software"), to deal in the Software without restriction,
@@ -17,30 +17,6 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-export interface ChangeProposalViewTabPanelProps {
-  changeProposal: ChangeProposal;
-}
-
-export interface ChangeProposal {
-  id: string;
-  name: string;
-  project: Project;
-}
-
-export interface Project {
-  identifier: string;
-  name: string;
-  organization: Organization;
-}
-
-export interface Organization {
-  identifier: string;
-  name: string;
-  role: MembershipRole;
-}
-
-export type MembershipRole = 'ADMIN' | 'MEMBER' | 'NONE';
-
-export interface ChangeProposalViewTabPanelState {
-  activeTab: number;
+export interface ProjectViewHeaderProps {
+  children?: React.ReactElement[];
 }
