+++ frontend/svalyn-studio-app/src/projects/settings/ProjectSettingsView.types.ts
@@ -18,48 +18,5 @@
  */
 
 export interface ProjectSettingsViewState {
-  name: string;
-  description: string;
   deleteProjectDialogOpen: boolean;
-}
-
-export interface UpdateProjectNameData {
-  updateProjectName: UpdateProjectNamePayload;
-}
-
-export interface UpdateProjectNamePayload {
-  __typename: string;
-}
-
-export interface UpdateProjectNameVariables {
-  input: UpdateProjectNameInput;
-}
-
-export interface UpdateProjectNameInput {
-  id: string;
-  projectIdentifier: string;
-  name: string;
-}
-
-export interface UpdateProjectDescriptionData {
-  updateProjectDescription: UpdateProjectDescriptionPayload;
-}
-
-export interface UpdateProjectDescriptionPayload {
-  __typename: string;
-}
-
-export interface UpdateProjectDescriptionVariables {
-  input: UpdateProjectDescriptionInput;
-}
-
-export interface UpdateProjectDescriptionInput {
-  id: string;
-  projectIdentifier: string;
-  description: string;
-}
-
-export interface ErrorPayload extends UpdateProjectDescriptionPayload, UpdateProjectNamePayload {
-  __typename: 'ErrorPayload';
-  message: string;
 }
